Fix some Wconversion 8/n (#2498)
*  Fix some Wconversion 8/n

* Fix padding checks.

+++ common/eager/pipeline_deleter.cpp
@@ -25,7 +25,7 @@ using namespace ::habana;
 
 namespace {
 
-int get_thread_tid() {
+long get_thread_tid() {
   return syscall(static_cast<long>(SYS_gettid));
 }
